adding missing references
We want to get them manually, because there are many references, that are not related to resources references and there are a lot of issues with that. If there will be any not working reference, we will need to fix it my providing correct path or name (example there is orgRef, and not organizationalRef - kind + Ref name.) (different path in cloudManagementRef - not in spec.ForProviders, but in spec.